Extracted classes StorageConfiguration, SecurityConfiguration
