+++ src/main/java/com/streamlined/tasks/repository/HashMapTrainingRepository.java
@@ -1,7 +1,6 @@
 package com.streamlined.tasks.repository;
 
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.stream.Stream;
